Now the csv file is being properly read and delimited to the right variables. Time to implement the simulation!

+++ game.h
@@ -40,7 +40,7 @@ private:
     bool isStartButtonClicked = false;
     std::string inputString;
     const char* textInput;
-    int maxHeight = 0.0f;
+    float maxHeight = 0.0f;
     float timeToLand = 0.0f;
     float timeToReachMaxHeight = 0.0f;
     bool hasStarted = false;
